Reorgnaised and added views
Reorganised the CSV files to better show where each CSV came from and
how it fits into the whole.

Updated the SQL with views, preserved the
foriegn keys on the reference data but commented them out as they
prevent cleaning the reference data to produce the cube.

Minor tweaks to the README